day6: main task finished
The solution feels very ugly. This is in part due to two things: One is
that input strings must be parsed into both instructions and a double
set of coordinates, which is a bit messy. The second is that 2D arrays
are not implemented in Rust, meaning that functions to manage that grid
must be done manually.

I ended up creating a somewhat simple system with Structs for both
individual coordinates and the general grid. I implemented an iterator
for slicing out a part of the grid and wrote a bunch of tests for it.
Messy stuff and can probably be improved tenfold, although the basic
concepts probably are fine.

One simple improvement would be to create a combined struct for the
light array (a boolean array) and the slicing operations. This would
resemble an object oriented style as seen in Python.