day6: restructure the code by implementing the array as a struct
To avoid stupid duplication of code and improve readability, the main
light array was moved to a struct `LightArray` which contains it and the
shape of the array as well as a constructor.

This struct implements a method to create a slice of given coordinates,
which as previously operates as an iterator. Code for advancing in the
coordinates was moved from the simple tuple to this slice, which knows
the shape of the full array and thus all hard-coded values are removed,
in addition to it making more sense.