day6: add error handling for Coordinate parsing
Use `filter_map` and `ok` to filter successful coordinate
transformations.

This should remove most of the remaining possible panic!'s from the
code. The big remaining one is if a Slice is created to go beyond the
limits of the LightArray. That would be easy enough to solve but I'm
pretty satisfied with leaving the exercise at this stage.